organizing diagnostics better
