+++ cmass/diagnostics/summ.py
@@ -8,36 +8,12 @@ import numpy as np
 import logging
 from os.path import join
 import hydra
-from omegaconf import DictConfig, OmegaConf
-import Pk_library as PKL
-import MAS_library as MASL
-import redshift_space_library as RSL
+from omegaconf import DictConfig
 import h5py
 
 from ..utils import get_source_path, timing_decorator
 from ..nbody.tools import parse_nbody_config
-
-
-def MA(pos, L, N, MAS='CIC'):
-    pos = np.ascontiguousarray(pos)
-    delta = np.zeros((N, N, N), dtype=np.float32)
-    MASL.MA(pos, delta, BoxSize=L, MAS=MAS)
-    delta /= np.mean(delta, dtype=np.float64)
-    delta -= 1
-    return delta
-
-
-def MAz(pos, vel, L, N, h, z, MAS='CIC', axis=0):
-    pos, vel = map(np.ascontiguousarray, (pos, vel))
-    RSL.pos_redshift_space(pos, vel, L, h*100, z, axis)
-    return MA(pos, L, N, MAS)
-
-
-def calcPk(delta, L, axis=0, MAS='CIC', threads=16):
-    Pk = PKL.Pk(delta, L, axis, MAS, threads, verbose=False)
-    k = Pk.k3D
-    Pk = Pk.Pk
-    return k, Pk
+from .tools import MA, MAz, calcPk
 
 
 def rho_summ(source_path, L, threads=16):
@@ -54,6 +30,7 @@ def rho_summ(source_path, L, threads=16):
     k, Pk = calcPk(rho, L, threads=threads)
 
     logging.info(f'Saving rho diagnostics to {outpath}')
+    os.makedirs(join(source_path, 'diag'), exist_ok=True)
     with h5py.File(outpath, 'w') as f:
         f.create_dataset('k', data=k)
         f.create_dataset('Pk', data=Pk)
@@ -86,6 +63,7 @@ def halo_summ(source_path, L, N, h, z, threads=16):
     hist, _ = np.histogram(hmass, bins=be)
 
     logging.info(f'Saving halo diagnostics to {outpath}')
+    os.makedirs(join(source_path, 'diag'), exist_ok=True)
     with h5py.File(outpath, 'w') as f:
         group = f.create_group('Pk')
         group.create_dataset('k', data=k)
@@ -120,6 +98,7 @@ def gal_summ(source_path, hod_seed, L, N, h, z, threads=16):
     kz, Pkz = calcPk(delta, L, MAS='NGP', threads=threads)
 
     logging.info(f'Saving gal diagnostics to {outpath}')
+    os.makedirs(join(source_path, 'diag'), exist_ok=True)
     with h5py.File(outpath, 'w') as f:
         group = f.create_group('Pk')
         group.create_dataset('k', data=k)
@@ -139,7 +118,6 @@ def main(cfg: DictConfig) -> None:
 
     cfg = parse_nbody_config(cfg)
     source_path = get_source_path(cfg, cfg.sim)
-    os.makedirs(join(source_path, 'diag'), exist_ok=True)
 
     threads = 16
